Move backend stuff into seperate project
