+++ DinoLauncherLib/Xdelta3.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace DinoLauncher;
+namespace DinoLauncherLib;
 
 public class Xdelta3 //: IDinoLauncher
 {
